requirements.txt
PySide6>=6.7
google-generativeai==0.8.2
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
pyttsx3==2.90
SpeechRecognition==3.10.4
pyaudio>=0.2.11; platform_system == "Windows"
sounddevice==0.5.0; platform_system != "Windows"
numpy==2.1.3
Pillow==11.0.0
python-dotenv==1.0.1

